W-15854291: Refactored code to collapse following hierarchy: BaseFileSystemOperations - FtpOperations
